Add starter article resource (#7)
Adds a basic implementation of the articles resource under a `Blog`
context. The following fields are present:

* `title`
* `description`
* `body`
* `favorites_count`
* `slug`

This commit also deletes `PageControllerTest` since we do not have
that file or route.

Note that `updated_at` and `created_ at` fail the Postman tests
because Elixir's `to_iso8601` return a different string than
JavaScript's `toISOString`.